requirements.txt
django==1.9.2
djangorestframework==3.3.2
djangorestframework-jsonapi==2.0.0-beta.1
django-filter==0.12.0
django-crispy-forms==1.6.0
psycopg2==2.6.1
coverage==4.0.3
mock==1.3.0
factory-boy==2.6.1
